LayoutTests/media/modern-media-controls/time-label/time-label.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>TimeLabel</code> class.");

const timeLabel = new TimeLabel;

shouldBe("timeLabel.value", "0");
shouldBeEqualToString("timeLabel.element.localName", "div");
shouldBeEqualToString("timeLabel.element.className", "time-label");

timeLabel.value = 1;

const timeLabelWithNegativeValue = new TimeLabel;
timeLabelWithNegativeValue.value = -61;

const timeLabelWithHours = new TimeLabel;
timeLabelWithHours.value = 3661;

let style;
scheduler.frameDidFire = function()
{
    document.body.appendChild(timeLabel.element);
    style = window.getComputedStyle(timeLabel.element);

    shouldBeEqualToString("style.position", "absolute");
    shouldBeEqualToString("style.fontFamily", "-apple-system-monospaced-numbers");
    shouldBeEqualToString("style.fontSize", "14px");

    timeLabel.element.remove();

    shouldBeEqualToString("timeLabel.element.textContent", "00:01");
    shouldBeEqualToString("timeLabelWithNegativeValue.element.textContent", "-01:01");
    shouldBeEqualToString("timeLabelWithHours.element.textContent", "01:01:01");

    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
