LayoutTests/media/modern-media-controls/time-label/time-label-expected.txt

Testing the TimeLabel class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timeLabel.value is 0
PASS timeLabel.element.localName is "div"
PASS timeLabel.element.className is "time-label"
PASS style.position is "absolute"
PASS style.fontFamily is "-apple-system-monospaced-numbers"
PASS style.fontSize is "14px"
PASS timeLabel.element.textContent is "00:01"
PASS timeLabelWithNegativeValue.element.textContent is "-01:01"
PASS timeLabelWithHours.element.textContent is "01:01:01"
PASS successfullyParsed is true

TEST COMPLETE
